fix(settings): Feed-Abos aus dem adminOnly-Blatt in die persoenliche Domaene
Beide Feed-Tokens haengen an der eigenen users-Zeile - calendar_feed_token
(Migration 61) und inventory_deadlines_feed_token (Migration 144) - und beide
Routen tragen serverseitig bewusst keinen Admin-Check. Erreichbar waren sie
trotzdem nur ueber Einstellungen > Sync > Kalender, und dieses Blatt ist
adminOnly: in einem fuenfkoepfigen Haushalt konnte genau eine Person ihr
eigenes Abo einrichten oder zurueckziehen, ohne dass eine Fehlermeldung das
erklaert haette - die Seite existierte fuer die anderen schlicht nicht.

Die zwei Abschnitte bekommen ein eigenes Blatt unter Persoenlich. CalDAV,
ICS-Abos und der Kalenderimport bleiben vorerst, wo sie sind.

Vierter Fall desselben Musters nach den Kalender-Vorgaben, den
Aufgaben-Vorgaben (#695) und der Navigationsreihenfolge; die Verteilungs-
zusicherung in test-settings-navigation.js zaehlt personal deshalb auf zehn.
