+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Inventory ships disabled and each household switches it on** - the first module with that default. Every module is a permanent line item in the navigation of every household, including the ones that will never track a bike, and Inventory is the first whose audience is visibly a subset. Enable it under Settings → Navigation; a household that does not want it never sees it, and existing installs do not find a new entry after an update they did not ask for.
 - The Inventory deadlines ICS feed token is now per user instead of household-wide, matching the calendar feed. A household-wide token could not be withdrawn from one person: whoever had subscribed kept access until it was taken from everyone. Each member now mints, rotates and disables their own subscription, and doing so leaves everyone else's running. The feed's *content* is unchanged and still household-wide - inventory items belong to the household, not to a member.
 - The Inventory list is a two-level browse: the landing page shows metric cards (item count, total purchase value, items needing attention) and a category overview, and tapping a category shows its items grouped by storage location with "All" / "Needs attention" filter chips scoped to that category. Items can carry a single photo, and tapping one opens a read-only detail view (with a colored accent stripe) before editing; the nav icon carries a badge for items with a soon-expiring or overdue warranty or tracked date. The "Needs attention" metric card on the landing page is itself a filter now - tapping it flattens the list across every category to just the items that need a look, the same view search results already use. The five seeded categories (Electronics, Vehicles, Household, Sports, Other) are localized instead of staying German regardless of the app's language, matching how Task Categories already handle their own seeded set.
+### Fixed
+
+- **Everyone can manage their own calendar and inventory feed subscription, not just the admin.** Both feed tokens have always been personal - they sit on your own user row, and neither route asks whether you are an admin - but the only place to reach them was Settings → Sync → Calendar, and that leaf is admin-only. In a household of five, one person could subscribe their phone to the family calendar and the other four could not, with no error to explain it: the page simply did not exist for them. The two sections now have their own leaf under **Settings → Personal → Feed subscriptions**, reachable by every member. Only the two feed sections moved; CalDAV accounts, ICS subscriptions and the one-time calendar import stay where they are for now. This is the fourth time a per-user setting had been stranded on an admin-only leaf, after the calendar defaults, the task defaults (#695) and the navigation order; existing subscription URLs keep working and nobody has to re-subscribe.
 
 ## [2.12.0] - 2026-08-15
