+++ public/locales/ar.json
@@ -2015,6 +2015,8 @@
         "pageCalendarDefaultsDescription": "حدد التذكيرات الافتراضية وإسناد المواعيد الجديدة الخاصة بك.",
         "pageNavigation": "التنقل",
         "pageNavigationDescription": "حدّد ترتيب الوحدات والتنقل على الهاتف.",
+        "pageFeeds": "Feed subscriptions",
+        "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
         "pageKitchen": "المطبخ",
         "pageKitchenDescription": "اختر الوجبات الظاهرة في خطة الوجبات.",
         "pageModuleOptions": "خيارات الوحدات",
