+++ public/locales/nl.json
@@ -2015,6 +2015,8 @@
         "pageCalendarDefaultsDescription": "Stel standaardherinneringen en toewijzing voor je eigen nieuwe afspraken in.",
         "pageNavigation": "Navigatie",
         "pageNavigationDescription": "De volgorde van de modules en de mobiele navigatie instellen.",
+        "pageFeeds": "Feed subscriptions",
+        "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
         "pageKitchen": "Keuken",
         "pageKitchenDescription": "Kies welke maaltijden zichtbaar zijn in het maaltijdplan.",
         "pageModuleOptions": "Module-opties",
